Create solution_dfs.py
O(nm) time and space solution using DFS and a visited set.

First, create the visited set.

Define the DFS function on indices (i, j) in the grid. For the base case, if the point (i, j) is out-of-bounds or the value at (i, j) in the grid is a 0 (water), then return 1 since a perimeter point is reached from the island. If (i, j) was already visited, return 0. Now, perform DFS on each of the adjacent cells in the grid from the current position (i +/- 1, j +/- 1), and add to the perimeter and return the perimeter.

Finally, go through the grid for each (i, j) combination in order to find a point that is land (value is 1), then start and return the DFS from that point.