Create solution.py
O(c) time solution, where c is the total number of characters in all of the words in the words list. Space is O(1) for storing the 26 lowercase alphabet ordering of with their corresponding indices.

For the index i until the 2nd-to-last word in the words list, get each 2 adjacent words w1 and w2 at index i and (i+1). Now, for each index in the length of the first word w1, if j equals the length of w2 (meaning w2 is a prefix of word w1), return false. If the characters at position j for w1 and w2 are different, if the order index dictionary has a lower value for w2 than w1, return false. Otherwise, break from the inner for loop to continue to the next 2 word pair iteration in the outer loop.

Finally, if the loop finishes without returning, return true.