Create solution_bfs.py
O(nm) time and space solution using iterative BFS, with n and m being the grid size.

First, if the grid is empty, return 0.

Get the row and column lengths.

Initialize the visited set to an empty set and the number of islands to 0.

Define the BFS function with input for the row and column. Create a queue, and add the (r, c) coordinate grid to the visited set and the queue. While the queue is not empty, pop from the queue to get the current row and column, and define the directions (right/left and up/down) deltas, and for each direction, increment the row and column by that amount. If the newly updated row/column pair is both in-bounds, is land (value is '1'), and not already visited, then append this (r, c) pair to the queue and to the visited set.

For each row/column in the grid, if the value is a '1' and has not been visited, run BFS on that row/column and increment the number of islands by 1.

Finally, return the number of islands.