Create solution_bfs.py
O(nm) time and space solution using iterative BFS, with n and m being the grid size.

First, if the grid is empty, return 0.

Get the row and column lengths.

Initialize the visited set to an empty set and the max island area to 0.

Define the BFS function with input for the row and column. Create a queue, and add the (r, c) coordinate grid to the visited set and the queue. Initialize the value of the area to the value at the grid at (r, c). While the queue is not empty, pop from the queue to get the current row and column, and define the directions (right/left and up/down) deltas, and for each direction, increment the row and column by that amount. If the newly updated row/column pair is both in-bounds, is land (value is 1), and not already visited, then append this (r, c) pair to the queue and to the visited set, and increment the area by 1.

For each row/column in the grid, if the value is a 1 and has not been visited, run BFS on that row/column and return the area. Find the max of this current island area with the running max island area.

Finally, return the max island area.