Create solution_dfs.py
O(mn) time and space solution using DFS and recursive stack.

If the grid is empty return 0.

Initialize the number of rows and columns, the visited set, and the max island area to 0.

Define the DFS function, with the base case of when either r or c is out-of-bounds, or the grid value at (r, c) is 0, or (r, c) is already in the visited set, then return 0. Otherwise, add (r, c) to the visited set, and return 1 + the DFS performed in all 4 direction +/- for each row/column pair.

For each row/column pair, loop through and if the value of the grid at that (r, c) coordinate is 1 and has not been visited yet, get the area using DFS and update the max island area.

Finally, return the max island area.