Create solution.py
O(n^2) time solution with O(1) space.

First, set the palindrome count to 0.

Define a count_palindromes() function taking in the string, and left and right pointers as input. Set the result to 0, then while the left and right pointers are inbound and the character at the left and right pointer positions in the string are equal, increment the result and widen the window by decrementing/incrementing the left/right pointers, respectively. Finally, return the result.

For i in the range of the length of the string s, first set the left and right pointers equal. This takes care of the odd-length palindromes. Increment the palindrome count by the value returned by the count_palindromes() function on the left and right pointers. Subsequently, change the left and right pointers to i and i + 1, respectively. This takes care of the even-length palindromes. Now, increment the palindrome count by the value returned by the count_palindromes() function on this updated set of left and right pointers.

Finally, return the value of the palindrome count.