Create solution.py
O(n^2) time solution and O(n) space solution, where n is the query row number.

First, get the previous row as the first (0-th) row with the total poured value. 

For each row in the range of 1-index-th row to the query_row, set the current row to a row of 0's with length row + 1. For each i in the row array, get the left and right partitions' values if the index i is in-bounds (otherwise set to 0). Then, get the extra amount of outflow for the left and right partitions, and calculate the current row's i-th value as the sum of half of the left and right extra overflows. After this inner loop, set the previous row to the current row.

Finally, return the minimum of 1 and the previous row's query glass index.