Create solution_opt.py
O(n^2) time and O(n) space solution, where n is the query row, optimized with only calculating the overflow up to each row.

First, get the previous row to be set to the initial 0-th row, with the poured total flow.

For each row from the 1-index-th row to the query row, set the current row to an array of 0's of size (row + 1). For each index up to the index equal to the current row, get the extra from the previous row's index - 1. If the extra is greater than 0, to each of the current row's i-th and (i+1)-th index, add half the extra overflow. Finally, after the inner loop, set the previous row to the current row.

Return the minimum of 1 and the previous row's value at the query glass.