Create solution_rev.py
O(nm) time solution with O(m) memory, where n is the number of words and m is the max (or average) size of a word in the words array.

Iterate over each word in the words array. If the word equals the reversed version of itself, return the word.

Otherwise, return the empty string after the iteration completes without returning anything.