Create solution.py
O(n) time and O(1) space-optimized solution using a dictionary storing only 2 elements at the most (since there is only 2 possible elements with count greater than n/3).

First, initialize a defaultdict of the counts.

For each number in the array, increment the value of that number key in the counts dictionary. If the length of the counts dictionary is less than or equal to 2, continue to the next iteration. Otherwise, create a new default dictionary, and for each key/value pair in the counts dictionary, only add the numbers whose counts are greater than 1, and set the count to be decremented by 1. Finally, assign the new count dictionary to the original counts dictionary.

Initialize a result array.

For each number in the counts dictionary (of which there should be no more than 2), count that number in the original array and check if it is greater than 1/3 of the length of the array, and if so, append to the result array.

Finally, return the result.