Create solution.py
O(n) time and O(1) space (not including output array) solution using prefix/postfix sums.

First, get the total sum of the nums array, and initialize the left sum to 0. 

Initialize the result array as empty.

Iterating through each index/number enumeration of the array, calculate the right sum by subtracting from the total sum the current number and the left sum. Then, calculate the size of the subarray to the left and right of the current element of the array. Then, append to the result array the absolute difference of subtracting from the current number (multiplied by cardinality of the left subarray) the left sum, and adding the right sum minus the cardinality of the right subarray multiplied by the current number. Finally, update the left sum by adding the current number.

Return the result array.