Create solution_ext.py
O(nm) time solution where n is the number of values in the nums array, and m is the max multiple of numbers to decode.

First, initialize the result array to an empty array. 

For i in range up to the end of the nums array in multiples of 2, extend the result array by the 2nd number of each number pair by the value of the 1st number.

Return the result.