Create solution.py
O(n) time and O(1) space solution using two pointers.

First, set the current index of the string and i index to 0.

While the index i is in-bounds, get the current character at the i-th index and set the count to 1. Now, while i + 1 is in-bounds and the next character (the character at index (i+1)) is equal to the current character, increment both the i and the count. Afterwards, set the character in the chars array at the current index to the current character, and increment the current index. Now, in order to add the run-length, if the count is greater than 1, iterate through the string version of the count integer, and set the character of the integer at the current index to its value, and increment the current index. Finally, in the outer loop, increment the index i.

Return the current index, which will be the length of the new array.