Create solution_bucketsort.py
O(m n log_2 n + m log m) time complexity, where m is the number of values in the array, and n is the number itself. Storage is also O(m) due to the overall buckets each containing the m values in some shape or form.

Create the buckets with size 15 (since 2^14 > 10000, the input value limit), with 0 needed. Initialize the result array.

For each number in the array, save the number to another variable, and set the number of one bits to 0. While the number is not 0, logical AND the number itself with the number minus 1, and increment the number of one bits. Then, append the original number to the bucket index with the number of corresponding one bits.

For each bucket in the buckets list, sort the bucket. Then, for each number in the bucket, append to the result array.

Finally, return the result.