Create solution_dict.py
O(n) time and O(52)=O(1) space solution using hashmap (dictionary).

First, create a default dictionary of counts. Initialize the result to 0.

For each character in the string, increment the count dict of that character. If the count dict value for the character is divisible by 2, then increment the result by 2 (since it is a pair - even palindrome).

For each count value in the count dictionary, if the value is odd, then increment the result by 1, then break out of the loop.

Return the result.