Create solution_hashset.py
O(n) time solution with O(1) space using a hashset.

First, create the seen characters set. Initialize the result to 0.

For each character in the string s, if it is already in the seen characters set, remove it and increment the result by 2 (this means an even number of this character was seen). Otherwise, add the character to the seen set.

Return result plus 1 if the seen set is not empty (since at least one odd number of characters exists), otherwise just return the result.