Create solution_dfs.py
O(n) time and space solution using recursive DFS.

First, create an old-to-new mapping for each of the graph nodes.

Define the DFS clone function for input node. If the node already exists in the old-to-new map, then just return the new node. Otherwise, create a copy node by creating a new node with the input node's value, and create the mapping of the old node to the copy node. For each neighbor of the old node, append to the copy node's neighbors the DFS result on the neighbor of the old node. Finally, return the copy node.

Return the DFS result on the node if the node is not null, otherwise return none.