Create solution_bfs.py
O(mn) time and space solution using simultaneous multi-source BFS with a queue.

First, get the dimensions of the rows and columns of the grid.
Initialize the visited set and the queue.

Define the addRooms function taking in the row and column. If the row or column is out-of-bounds, or (r, c) coordinate has already been visited or the value of the grid at that point is a wall (-1), then return. Otherwise, add this (r, c) coordinate pair to the visited set and the queue.

For each r/c pair in the rows and columns, check if the value is a 0 (a gate). Then append this (r, c) coordinate to the queue and the visited set.

Set the distance to 0.

While the queue is not empty, for each value in the queue pop and get the coordinate. Then, add the rooms for each +/- row and column indices. After this inner loop, increment the distance by 1.

Nothing needs to be returned as the grid modification is done in-place.