Create solution_bfs.py
O(mn) time and space solution using BFS.

First, get the row and column dimensions of the grid. Initialize the queue, number of fresh oranges, and time to 0.

For each row/column pair in the grid, if the grid value is 1, increment the number of fresh oranges by 1. If the value is 2 (a rotten orange), then append the coordinates to the queue.

Create the directions delta list for each of the 4 possible directions of travel.

While the queue is not empty and the number of fresh oranges is greater than 0, for each value in the queue, pop and get the coordinates. Now, for each delta in the row/column directions (+/- 1), update the row and column. If they are out of bounds or the value is not 1 (so not a fresh orange), continue to the next iteration. Otherwise, set the value at that grid point to 2 (rotten), append that coordinate to the queue, and decrease the number of fresh oranges by 1. After the two inner loops, increment the time by 1.

Finally, return the time if the number of fresh oranges is 0, otherwise return -1.