Create solution_dfs.py
O(mn) time and space solution using DFS.

First, get the row and column dimensions and define the visited set.

Define the DFS function taking in the row and column indices. If the row or column is out-of-bounds, or the value in grid2 at (r, c) is 0 (water) or (r, c) is already visited, then return true. Otherwise, add (r, c) to the visited set and set result to true. If the grid1 value at (r, c) is 0, then set the result to false. Now, set the result to DFS in one direction and the result itself for each of the possible 4 directions. The order here matters; DFS function calls logical ANDed with res is the correct order (other way gives incorrect answer). Finally, return the result.

Set the count to 0.

For each row/column pair iteration, if the value at grid2 at (r, c) is 1 and it has not been visited, and the DFS at (r, c) is true, then increment the count by 1.

Finally, return the count.