Create solution_dfs.py
O(mn) time solution with same space using recursive DFS.

First, get the dimensions of the grid.

Initialize the atlantic and pacific sets, which show the cells reachable from the atlantic and pacific oceans, respectively.

Define the DFS function with row and column inputs, a visited set passed in, and the previous cell's height. If the row or column is out-of-bounds, or if the coordinate (r, c) has already been visited, or the current height is less than the previous height, return. Otherwise, add (r, c) to the visited set and perform DFS in all 4 directions.

For each column, perform DFS on the first row (which is the pacific) and last row (which is the atlantic).

Similarly, for each row, perform DFS on the first column (which is the pacific) and last column (which is the atlantic).

Initialize the result array.

Now, go through the grid once more, and if the coordinate (r, c) is in both the atlantic and pacific visited sets, append the list form [r, c] of the coordinates to the result list.

Return the result.