Create solution.py
O(n) time and O(1) space solution by finding a pivot point and reversing portion (or whole) of the array in-place.

For the array starting from the right, if the current value in the array is greater than the value to its left (i-1 index), then set that to the pivot index and break out. If the loop finishes without any breaks, then this means the array is sorted in reverse order, so just reverse the whole array and return.

Set the swap index to the end of the array. 

While the value at the swap index is less than or equal to the pivot value, keep decrementing the swap index.

Then, swap the values at the pivot index and the swap index.

Finally, reverse the portion of the array to the right of the pivot index.

Return out of the function.