Create solution_dfs.py
O(nm) time and space solution, where n and m are the dimensions of the grid, with in-place modification of original grid.

First, get the dimensions of the grid.

Define the DFS function. If r or c is out-of-bounds, or the value at the board is not 'O', then return. Otherwise, set that value of 'O' connected to the outside border to a temporary 'T', then perform DFS in all 4 adjacent directions.

1) (convert 'O' on border edges to 'T'): For r in rows, perform DFS on the leftmost and rightmost columns, and similarly, perform DFS on the topmost and bottommost rows. This converts all 'O's reachable from the edge of the grid to 'T' since they will never be surrounded.

2) (convert remaining 'O' which are surrounded to 'X'): For each (r, c), if there is still an 'O', convert them to an 'X' since they are surrounded.

3) (convert 'T' to 'O'): For the remaining 'T', similarly iterate through the whole grid and convert them to an 'O'.