Create solution_dfs.py
O(n) time and space solution using DFS recursively.

First, create an edges set for each source-sink pair in connections. Also, create a hashmap of city to empty lists.

Initialize a visited set and set the number of changes needed to 0.

For each source/sink pair a/b in connections, create mutual neighbors between a and b both directions (just to indicate neighborship).

Define the DFS function with an integer city number input. Indicate that the non-object integer value num_changes is nonlocal. Then, for each neighbor of the current city, if the neighbor was already visited, continue to the next neighbor. Otherwise, if the neighbor-city pair is not in the edges, then increment the number of changes by 1, and add the neighbor to the visited set and perform DFS on the neighbor.

Add the capital city 0 to the visited set, and perform DFS starting from this city.

Finally, return the number of changes.