Create solution_bfs.py
O(n^2) time and space solution using BFS, where n is the length of one side of the square board.

First, get the length of the board.

Reverse the board (for ease in calculating the integer value to board coordinate calculation function).

Define the integer to coordinate converter function. First, get the row index by subtracting 1 from the number then doing integer division by the board length. Similarly, get the column index by subtracting 1 and modding by the board length. Now, if the row is odd, then the direction of the numbers are reversed, so in order to swap the column index, subtract from (n - 1) the column index. Finally, return the row-column coordinate pair.

Initialize a queue that will contain tuples of the square number and moves performed so far. Append the first square (1) with value of moves 0.

Initialize the visited set.

While the queue is not empty, get the square number and moves performed by popping from the queue. Then, for the possible 6 moves of the dice (1-6), calculate the next square position. Convert the next square to the appropriate coordinates. If the value at the board coordinate (r, c) is not -1, then it is a snake or ladder, so update the next square position to that value. If the next square afterwards happens to be n^2, then return the moves plus 1 (which is the move just completed), and it is done. Otherwise, as long as the square was not visited, add to the visited set, and append this square and number of moves + 1 to the queue.

Finally, if the loop finishes without returning, return -1.