Create solution_bfs.py
O(10^4) = O(1) time and space solution for all 10000 possible lock combinations using BFS.

First, check if the initial position '0000' is a deadend. If so, return -1.

Initialize the queue with a tuple of the start position of the lock '0000' with number of turns equal to 0.

Initialize the visited set with the deadends (as if they were already visited).

Define the children function to get the children of a lock position (possible lock positions 1 turn away). Initialize the result array. For each i in the 4 positions of the lock, first calculate turning the lock to increase the value. Mod the value by 10, then convert to a string and append to the result. Similarly, subtract 1 from the position at index i for decrementing the lock position, then add 10 and mod by 10. Append to the result after only changing that position and constructing the string. Return the result.

While the queue is not empty, get the current lock position and the number of turns to reach the position so far. If the lock position equals the target, return the number of turns. For each child in the children of the current lock position, if it is not in the visited set, mark it as visited and append the child and number of turns + 1 to the queue.

Finally, if the queue exits, return -1.