Create solution.py
O(n) time and O(n) space solution.

First, initialize the result array.
Set the open parenthesis surplus count to 0.

For each character in the string, if the it is an open parentheses '(', then increment the count by 1 and append to the result array. If the character is a closing parenthesis ')' and the surplus count is greater than 0, then append the character to the result array and decrement the count. If the character otherwise is not a closing parenthesis, append the character to the result.

Create an updated result array.

For the characters in the reversed result array, if the character is '(' and the count is greater than 0, decrement the count. Otherwise, append the character to the updated result array.

Finally, return the joined string version of the reversed updated result array.