Create solution.py
O(n log n) time and O(1) space solution using sorting and iterating through the array.

First, sort the array.

Initialize the result to -1, and the current running sum to 0.

For each number in the sorted array, if the current sum is greater than the current number, update the result. The if statement does not need to check if i is greater than or equal to 2 (3 sides at least for the polygon), since there are no cases where the current sum will be greater than the current number while i < 2. Afterwards, add the current number to the current sum.

Finally, return the result.