Create solution.py
O(E + V) time and space solution using DFS, with E being the number of edges in the graph and V being the number of vertices (nodes).

First, get the length of the graph.

Create the node safety hashmap, which maps whether each node is safe.

Initialize the result array.

Define the DFS function on a node i. If i is already in the node safety map, return the value in the map. Otherwise, by default set the node safety to false. For each neighbor of node i, if DFS of the neighbor is false, return its value in the node safety map (which will always be false by default). If the loop does not return, then set the node safety at i to true and return it.

For each node index in the range, if the DFS at i is true, append the i to the result array.

Finally, return the result.