Create solution.py
O(mn) time and space solution using BFS, where m is the number of buses and n is the max number of stops at a bus station.

First, if the source equals the target, return 0.

Create a default dict of sets to create a graph of the stops corresponding to a set of buses that stop at that stop.
Similarly, initialize a queue with the source node, and 0 representing the route length measure by number of buses.

For each bus/route pair in routes, for each stop in the route, add to the graph at each stop the corresponding bus.

Create a visited set for the stops visited and the buses visited.

While the queue is not empty, get the stop and the route length so far by popping from the queue.

If the stop is the target, return the route length. For each bus in the graph's stop, if the bus has not been visited, add it to the bus visited set, and for each stop in the routes for that bus, if the stop has not been visited, add the stop to the stop visited set, and append the stop to the queue with the route length incremented by 1.

If the loop does not return, return -1.