Create solution_dp_recur_cache.py
O(n) time and O(n) space solution using recursive DFS for DP and caching to DP array.

First, initialize the DP array with the last index (length of the string) equal to 1 (since an empty string also returns value 1).

Define the DFS on index i of the string. If i is already in the DP map, return the value. If the character at index i is '0', return 0. Set the result to DFS on the next index i + 1. If (i + 1) is less than the length of the string and the character at index i is a 1 or is a 2 and the i + 1 index is between 0-6 (essentially checking if the double-digit value is between 10-26), then increment the result by DFS on i + 2. Update the DP array with the result, and return the result.

Return the result from running DFS at index 0.