Create solution.py
O(nm) time and O(1) space solution, where n is the number of rows and m is the number of columns in the strs array.

First, get the dimensions of the rows (number of strings in the list) and the columns (the length of each string).

Set the number of columns to 0.

For each column, first set the current character to 'a' (since the strings only conists of lowercase English alphabet letters). Then, go through each row. If the character at point (r, c) is less than the current character, increment the number of columns to remove by 1 and break out of this inner loop. Otherwise, set the current character to the character at (r, c) in the list of lists.

Finally, return the number of columns.