Create solution.py
O(n) time and O(1) space solution.

First, get the max and min values of the array (O(n) operations). 

Then, for each number in the array, if the number if neither equal to the maximum nor the minimum values, return the value.

Otherwise, if the loop finishes without returning, return -1.