Create solution.py
O(n) time solution for each character with O(26) = O(1) auxiliary space.

First, create a pangram indicator array of size 26 for the 26 lowercase alphabet. Initialize the character sum to 0.

For each character in the sentence, calculate the ordinal index with respect to the lowercase 'a'. If the value at that index in the pangram indicator array is 0, then set it to 1 and increment the character sum. If the sum reaches 26, return true. Otherwise, continue.

Return whether the character sum is 26 or not.