Create solution.py
O(n) time solution with O(1) space (not counting result array).

Initialize the result array.

For each i in range 1 to n, create a new empty string s. If the value i is divisble by 3, concatenate 'Fizz'. If the value is divisible by 5, concatenate 'Buzz'. If the string s is still an empty string, concatenate the string value of i. Append the string to the result array.

Finally, return the result array.