Create solution.py
O(n log n) time and O(n) space solution.

First, sort the even-index values by getting the values of each index divisible by 2 then sorting. Similarly, sort the odd-index values by getting the values of each index not divisible by 2, then sorting in reverse order.

Initialize the result array.

For the length of the odd-index values (which will always be less than or equal to the length of the even-index values), append the even-index list and then the odd-index list values, respectively, to the result array.

If the length of the even-index values is greater than the odd-index values, then append to the result the last element of the even-index values.

Return the result.