Create solution.py
O(n) time solution with O(1) space using two pointers and sorting by parity in-place.

First, set the left and right pointers to the ends of the list.

While the left pointer does not go out-of-bounds of the end of the array and while the right pointer does not go out-of-bounds of the beginning of the array, if the value at the left pointer is divisible by 2, increment the left pointer by 2. Otherwise, if the value at the right pointer is not divisible by 2, then decrement the right pointer by 2. Otherwise, swap the left and right pointer values, then increment/decrement the left/right pointers by 2, respectively.

Finally, return the array.