Create solution_dict.py
O(n log n) time and O(n) space solution using arrays and dictionaries to save the even and odd digits in sorted order, as well as keeping the ordering.

First, convert the number into a string. Iterate through the enumerated portion and get the indices for the even digits, and sort both the numbers themselves (in decreasing order) and indices (in increasing order). Create a map of the index-to-value map using these two lists. Perform this for the odd numbers as well.

Finally, merge the two odd and even index-to-value dictionaries.

Initialize a result array with -1 of length equal to the number string.

For each index in the string length, set the result at that index equal to the value in the dictionary for that key.

Finally, return the integer form of the joined result string.