Create solution.py
O(n) time and space solution using a dictionary.

First, create a default dictionary of lists for each value encountered.

For each index/number pair in the array, set the key as the number and append to the value list the index.

Set the degree to -1 and the smallest length to be 1 + the length of the nums array.

For each key-value pair in the count dictionary, if the length of the list is greater than the degree, set the degree to the length of the indices list, and the smallest length required to the last value in the list - the first value in the list + 1 (r - l + 1). Otherwise, if the length of the index list equals the degree but the index length in the original array (r - l + 1) is smaller than the current smallest length, then update the smallest length.

Finally, return the smallest length.