Create solution.py
O(n) time and space solution using a stack.

First, initialize the stack.

For each character in the string, if the stack is not empty and the top of the stack is  'C' and the current (next) character is a 'D', pop from the stack. If similarly the stack is not empty and the top of the stack is 'A' and the current (next) character is a 'B', pop from the stack. Otherwise, append the character to the stack.

Return the length of the stack.