Create solution.py
O(n) time and space solution using dictionaries.

First, count the number of occurrences of the values in the array and put a dictionary of this.
Create an array index to value dictionary of the enumerated array.

For each key in the array index dictionary (which will strictly have more than or equal to the number of keys in the count dictionary), if the key k does not exist in count dictionary but its value is equal to 0 in the array index dictionary, or if the k is in the count dictionary and the values of the corresponding key k is equal in both the count dictionary and the array index dictionary, then continue to the next key iteration. Otherwise, they are not equal, and return false.

Finally, if the loop ends without returning, return true.