Create solution.py
O(n log n m log m) time solution where n is the number of tables and m is the number of food items possible, with O(nm + n + m) storage for the result and dictionaries and sets.

First, create the table number to food default dictionary of lists which will store all the items for the table.

Initialize the food items and table numbers sets, and the result array.

For each triple of the customer name, table, and food item, append the food item to the corresponding table number in the  table to food dictionary, and add the table number and food item to the respective sets.

Now, iterating through the sorted table numbers set (in increasing order), get the food count dictionary with the list value in the table to food dictionary using a counter. Also create a food count list. For each food item in the sorted order (alphabetically), extend the food count with the string of the value in the food count dictionary. After the inner loop, append to the result array the table number list extended with the food count list.

Sort the result array by the key of the first value (the table number).

Return the first row of the 'Table' and sorted food items list as the first row and the result list concatenated afterwards.