Create solution.py
O(n) time and space solution.

First, create the result list as the list of the original string's characters.

For index i from 1 to the end of the string length, get the previous and current values as integers for the (i-1)-th and i-th indices. If the current value is less than the previous value and they have the same parity, swap them and break out (since there is only 1 swap allowed between adjacent digits).

Finally, return the joined string of the result list.