Create solution_gt.py
O(n^2) time and O(n) space solution,  where n is the length of the string.

This is a group theory solution by finding the 2 operations of adding 'a' to the odd indices and rotating by 'b' can be the composition of each of these operations.

Both operations are cyclic, i.e., ends up to the original string s if performed 10 times each for operation a, and ends up to the original string if performed n times for operation b (where n is the length of the string).

Since the periods of each operation will be 10 / gcd(10, a) for operation (a) and n / gcd(n, b) for operation (b), this cycle will optimize runtime.

When b is even, the group is commutative, so perform all (a) operations first then (b) operations (since order doesn't matter). Similarly, if b is odd, add the operations for adding 'a' to even indices, which makes all 3 of these operations commutative.

Define helper functions for all 3 operations, and then perform the operation the specified number of times, then return the result.