Create solution.py
O(n) time and O(n) space using dictionaries.

First, get the split string word counter and keysets for both s1 and s2.

Get the difference keyset between the s1 and s2 keysets.

Initialize the result array.

For each key in the keyset, if the key k is in the word count 1 dictionary and its value is 1, or if the key k is in the word count 2 dictionary and its value is 1, append this key to the result. Otherwise, continue.

Finally, return the result.