Create solution.py
O(nm) time solution, where n is the number of rows and m is the number of columns of the image grid.

For each row in the image, do a bitwise XOR with 1 for each element, then reverse the row using [::-1} (not reverse(), since it is slower). The ordering of inverting bits and flipping does not matter (group theory), they are commutative.

Return the image.