Create solution.py
O(nm) time solution with O(nm) space for a 1D array flattened for the 2D grid representation.

First, calculate the offset of the column characters using ord('A').

Calculate the integer values for the rows, then calculate the ord of the characters of the columns minus the offset.

Create an array of empty strings of the size of the grid.

For each column in the range from c1 to c2 (inclusive), go through each row from r1 to r2 inclusive, and calculate the index of the row by offsetting the row number by r1, and adding to it the column offset times the length of the row. Populate that value with the character corresponding to the column, concatenated with the row represented by that row.

Return the result.