Create solution.py
O(n) time solution going through the words list length and string length, where n is the shorter length of either the length of the words list or the string. O(1) space from pointer space.

First, initialize two pointers, p1 and p2 to 0.

While p1 is not out-of-bounds of the size of the words list and p2 is not out-of-bounds of the size of the string s, and the first character of the word in words list at pointer p1 and the string character at pointer position p2 are equal, keep incrementing p1 and p2.

Return whether the length of words, p2, and s are equal.