Create solution.py
O(n) time and O(n) space complexity, where n is the length of the string. The string is at most 26 characters long, so technically, it is O(1) time and space complexity.

Create a dictionary of character-to-index mapping for both strings s and t.

Populate the dictionary with the key as the character and the value as the index for both the s and t strings.

Initialize the result to 0.

For each key k in the s dictionary, add to the result the absolute difference.

Return the result.