Create solution_tabulation.py
O(n) time and O(n) space solution using tabulation.

First, if n is 0 or 1, return n.

Create a table of size (n + 1) initialized with -1. Set the 0-th and 1-st indices to 0 and 1, respectively.

For each index from 2 to n, add the previous two indices and set to its value.

Return the value at position n of the array.