Create solution_str_arr.py
O(n + m) time solution with O(n) space, where n is the size of the string to be rotated and m is the size of the shift array.

First, define the left_shift() and right_shift() functions, which take a string array (list) and shift amount. Get the length of the string array, and mod the shift amount by the value. Then, for the range of the shift amount, get the first character (or last character) for the left/right shift functions, and iterate (forwards or backwards), modifying the current value in the string array to be the next (or previous) character. Finally, adjust the first/last character, then join the string array into a string and return it.

Calculate the modulus size, set the initial shift amount to 0. Now, for each direction/amount in the shift array, increment for direction 0 (left-shift) as a convention, and decrement for direction 1 (right-shift) for the shift amount variable. Finally, mod this value by the length of the string, and if the value is positive (left-shift), pass it to the left shift function. If the value is negative, negate and pass it to the right shift function.

Return the string afterwards (the case where shift amount is 0, the string s is the same).