Create solution_deque.py
O(n + m) time and O(n) space solution using a deque.

First, define the shiftByAmount function to rotate the deque by the shift amount (negative is left rotation, positive is right rotation). 

Then, get the modulus size as the length of the string, and the shift number. For the direction and amount in the shift array, if the direction is 0 (left), decrement, if the direction is 1 (right), increment. Mod the shift number by the string size. 

Set the string to the shifted string if shift number is not 0; otherwise, leave it as-is.

Return s.